move list-related code into list.*
