+++ src/model/list.h
@@ -18,11 +18,12 @@
 #ifndef GOLOGPP_LIST_EXPRESSION_H_
 #define GOLOGPP_LIST_EXPRESSION_H_
 
-#include "language.h"
 #include "gologpp.h"
+#include "language.h"
 #include "expressions.h"
 #include "scope.h"
 #include "types.h"
+#include "semantics.h"
 
 #include <unordered_map>
 
@@ -52,6 +53,107 @@ private:
 	vector<unique_ptr<Expression>> entries_;
 };
 
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+class ListAccess
+: public Expression
+, public NoScopeOwner
+, public LanguageElement<ListAccess>
+{
+public:
+	ListAccess(Expression *subject, Expression *index);
+	const Expression &subject() const;
+	const Expression &index() const;
+
+	DEFINE_ATTACH_SEMANTICS_WITH_MEMBERS(*subject_, *index_)
+
+	virtual const Type &type() const override;
+
+	string to_string(const string &pfx) const override;
+
+private:
+	SafeExprOwner<ListType> subject_;
+	SafeExprOwner<NumberType> index_;
+};
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+class ListLength
+: public Expression
+, public NoScopeOwner
+, public LanguageElement<ListLength, NumberType>
+{
+public:
+	ListLength(Expression *subject);
+	const Expression &subject() const;
+
+	DEFINE_ATTACH_SEMANTICS_WITH_MEMBERS(*subject_)
+
+	string to_string(const string &pfx) const override;
+
+private:
+	SafeExprOwner<ListType> subject_;
+};
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+enum ListOpEnd {
+	FRONT, BACK
+};
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+class ListPop
+: public Instruction
+, public NoScopeOwner
+, public LanguageElement<ListPop, VoidType>
+{
+public:
+	ListPop(Expression *list, ListOpEnd which_end);
+	const Expression &list() const;
+	ListOpEnd which_end() const;
+
+	DEFINE_ATTACH_SEMANTICS_WITH_MEMBERS(*list_)
+
+	string to_string(const string &pfx) const override;
+
+private:
+	SafeExprOwner<ListType> list_;
+	ListOpEnd which_end_;
+};
+
+/////////////////////////////////////////////////////////////////////////////////////////////////
+/***********************************************************************************************/
+/////////////////////////////////////////////////////////////////////////////////////////////////
+
+class ListPush
+: public Instruction
+, public NoScopeOwner
+, public LanguageElement<ListPush, VoidType>
+{
+public:
+	ListPush(Expression *list, ListOpEnd which_end, Expression *what);
+	const Expression &list() const;
+	ListOpEnd which_end() const;
+	const Expression &what() const;
+
+	DEFINE_ATTACH_SEMANTICS_WITH_MEMBERS(*list_, *what_)
+
+	string to_string(const string &pfx) const override;
+
+private:
+	SafeExprOwner<ListType> list_;
+	ListOpEnd which_end_;
+	unique_ptr<Expression> what_;
+};
 
 
 } // namespace gologpp
