+++ src/model/procedural.cpp
@@ -597,111 +597,6 @@ string FieldAccess::to_string(const string &pfx) const
 /***********************************************************************************************/
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
-ListAccess::ListAccess(Expression *subject, Expression *index)
-: subject_(subject)
-, index_(index)
-{
-	subject_->set_parent(this);
-	index_->set_parent(this);
-}
-
-const Expression &ListAccess::subject() const
-{ return *subject_; }
-
-const Expression &ListAccess::index() const
-{ return *index_; }
-
-
-const Type &ListAccess::type() const
-{
-	return dynamic_cast<const ListType &>(
-		subject_->type()
-	).element_type();
-}
-
-string ListAccess::to_string(const string &pfx) const
-{ return subject_->to_string(pfx) + '[' + index_->str() + ']'; }
-
-/////////////////////////////////////////////////////////////////////////////////////////////////
-/***********************************************************************************************/
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
-ListLength::ListLength(Expression *subject)
-: subject_(subject)
-{ subject_->set_parent(this); }
-
-const Expression &ListLength::subject() const
-{ return *subject_; }
-
-string ListLength::to_string(const string &pfx) const
-{ return "length(" + subject_->to_string(pfx) + ')'; }
-
-
-
-string to_string(ListOpEnd which_end)
-{
-	switch (which_end) {
-	case FRONT:
-		return "front";
-	case BACK:
-		return "back";
-	}
-	throw Bug(string("Unhandled ") + typeid(which_end).name());
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////
-/***********************************************************************************************/
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
-ListPop::ListPop(Expression *list, ListOpEnd which_end)
-: list_(list)
-, which_end_(which_end)
-{ list_->set_parent(this); }
-
-const Expression &ListPop::list() const
-{ return *list_; }
-
-ListOpEnd ListPop::which_end() const
-{ return which_end_; }
-
-string ListPop::to_string(const string &pfx) const
-{ return pfx + "pop_" + gologpp::to_string(which_end_) + '(' + list_->str() + ')'; }
-
-/////////////////////////////////////////////////////////////////////////////////////////////////
-/***********************************************************************************************/
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
-ListPush::ListPush(Expression *list, ListOpEnd which_end, Expression *what)
-: list_(list)
-, which_end_(which_end)
-{
-	list_->set_parent(this);
-
-	what->ensure_type(
-		dynamic_cast<const ListType &>(
-			list_->type()
-		).element_type()
-	);
-	what_.reset(what);
-	what_->set_parent(this);
-}
-
-const Expression &ListPush::list() const
-{ return *list_; }
-
-ListOpEnd ListPush::which_end() const
-{ return which_end_; }
-
-const Expression &ListPush::what() const
-{ return *what_; }
-
-string ListPush::to_string(const string &pfx) const
-{ return pfx + "push_" + gologpp::to_string(which_end_) + '(' + list_->str() + ')'; }
-
-/////////////////////////////////////////////////////////////////////////////////////////////////
-/***********************************************************************************************/
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
 During::During(
 	Reference<Action> *action_call,
 	Instruction *parallel_block,
@@ -749,7 +644,6 @@ string During::to_string(const string &pfx) const
 		+ "on_cancel " + on_cancel().to_string(pfx)
 	;
 }
-
