+++ src/parser/action.cpp
@@ -40,6 +40,7 @@
 
 #include <model/action.h>
 #include <model/effect_axiom.h>
+#include <model/list.h>
 
 #include <boost/optional/optional_io.hpp>
