+++ src/parser/arithmetic.cpp
@@ -18,7 +18,7 @@
 #include "arithmetic.h"
 #include "variable.h"
 #include "value.h"
-#include "list_expression.h"
+#include "list.h"
 #include "mixed_member_access.h"
 #include "expressions.h"
 #include "reference.h"
