+++ src/semantics/readylog/history.cpp
@@ -66,6 +66,9 @@ shared_ptr<Transition> Semantics<History>::get_last_transition()
 void Semantics<History>::append(const AbstractEvent &trans)
 { extend_history(::list(trans.semantics().plterm(), plterm())); }
 
+void Semantics<History>::append(shared_ptr<AbstractEvent> e)
+{ append(*e); }
+
 
 void Semantics<History>::append_sensing_result(
 	shared_ptr<Activity>,
@@ -121,6 +124,7 @@ void Semantics<History>::progress()
 
 	readylog_history_ = New_history;
 }
+
