+++ src/semantics/readylog/list.cpp
@@ -37,4 +37,81 @@ EC_word Semantics<ListExpression>::plterm()
 }
 
 
+
+EC_word Semantics<ListAccess>::pl_index()
+{ return element().index().semantics().plterm(); }
+
+
+EC_word Semantics<ListAccess>::plterm()
+{
+	return ::term(EC_functor("gpp_list_access", 2),
+		element().subject().semantics().plterm(),
+		pl_index()
+	);
+}
+
+
+
+template<>
+EC_word Semantics<ListPop>::plterm()
+{
+	string fn;
+	switch(element().which_end()) {
+	case ListOpEnd::BACK:
+		fn = "gpp_list_pop_back";
+		break;
+	case ListOpEnd::FRONT:
+		fn = "gpp_list_pop_front";
+	}
+
+	if (fn.empty())
+		throw Bug("Invalid ListOpEnd Enum value: " + std::to_string(element().which_end()));
+
+	return ::term(EC_functor("set", 2),
+		element().list().semantics().plterm(),
+		::term(EC_functor(fn.c_str(), 1),
+			element().list().semantics().plterm()
+		)
+	);
+
+}
+
+
+
+template<>
+EC_word Semantics<ListPush>::plterm()
+{
+	string fn;
+	switch(element().which_end()) {
+	case ListOpEnd::BACK:
+		fn = "gpp_list_push_back";
+		break;
+	case ListOpEnd::FRONT:
+		fn = "gpp_list_push_front";
+	}
+
+	if (fn.empty())
+		throw Bug("Invalid ListOpEnd Enum value: " + std::to_string(element().which_end()));
+
+	return ::term(EC_functor("set", 2),
+		element().list().semantics().plterm(),
+		::term(EC_functor(fn.c_str(), 2),
+			element().list().semantics().plterm(),
+			element().what().semantics().plterm()
+		)
+	);
+}
+
+
+
+template<>
+EC_word Semantics<ListLength>::plterm()
+{
+	return ::term(EC_functor("gpp_list_length", 1),
+		element().subject().semantics().plterm()
+	);
+}
+
+
+
 } // namespace gologpp
