+++ src/semantics/readylog/procedural.h
@@ -132,20 +132,6 @@ public:
 
 
 
-template<>
-class Semantics<ListAccess>
-: public Semantics<Expression>
-, public GeneralSemantics<ListAccess>
-{
-public:
-	using GeneralSemantics<ListAccess>::GeneralSemantics;
-
-	virtual EC_word plterm() override;
-	EC_word pl_index();
-};
-
-
-
 } // namespace gologpp
